Support for multiple unity locations confoming to "/Application/Unity*" pattern. Check of already installed.

+++ macos_install.sh
@@ -1,13 +1,22 @@
 #!/bin/bash
+for unity_dir in /Applications/Unity* 
+do
+	ORIGINAL_LOCATION="${unity_dir}/Unity.app/Contents/Tools/PVRTexTool"
+	COPY_OF_ORIGINAL_LOCATION="${unity_dir}/Unity.app/Contents/Tools/PVRTexTool_orig"
 
-ORIGINAL_LOCATION=/Applications/Unity/Unity.app/Contents/Tools/PVRTexTool
-if [ -f $ORIGINAL_LOCATION ]; then
-	echo "Creating a copy of the original"
-	cp ${ORIGINAL_LOCATION} ${ORIGINAL_LOCATION}_orig
-	echo "Copying wrapper"
-	cp PVRTexTool $ORIGINAL_LOCATION
-	echo "Done"
-else
-	echo "Unity PVRTexTool not found. Doing nothing."
-	exit 1
-fi
+	echo "Installing into ${unity_dir}"
+
+	if [ -f $COPY_OF_ORIGINAL_LOCATION ]; then
+		echo "Already installed here."
+	else
+		if [ -f $ORIGINAL_LOCATION ]; then
+			echo "Creating a copy of the original"
+			cp ${ORIGINAL_LOCATION} ${COPY_OF_ORIGINAL_LOCATION}
+			echo "Copying wrapper to ${COPY_OF_ORIGINAL_LOCATION}"
+			cp PVRTexTool $ORIGINAL_LOCATION
+			echo "Done"
+		else
+			echo "Unity PVRTexTool not found. Doing nothing."
+		fi
+	fi
+done
